feat: implement interval jitter
Adds a new argument called 'interval_jitter', expressed as a percentage
number (range 0-100) that has the following effects:

If 0 => no effect.

If >0 => choose a random integer in the
[-(interval * jitter / 100), +(interval * jitter / 100)] range

and add use it as an offset for the sleep interval.

So for example, if our interval is 1000ms and we pick
interval_jitter=20, then the sleep values will be
in the range [800ms, 1200ms]

This has been a requested feature in the original implementation,
and is useful to twarth SSH scanners that have tarpit-detecting logic.

Link: https://github.com/skeeto/endlessh/issues/71